Moved FunctionHelpersXL.java to a separate package
