+++ .idea/checkstyle-idea.xml
@@ -3,12 +3,12 @@
   <component name="CheckStyle-IDEA">
     <option name="configuration">
       <map>
-        <entry key="active-configuration" value="LOCAL_FILE:$PROJECT_DIR$/checkstyle.xml:Project Checkstyle" />
+        <entry key="active-configuration-0" value="LOCAL_FILE:/home/idans/IdeaProjects/dremio-udf-gis/checkstyle.xml:Project Checkstyle" />
         <entry key="checkstyle-version" value="9.3" />
         <entry key="copy-libs" value="false" />
-        <entry key="location-0" value="BUNDLED:(bundled):Sun Checks" />
-        <entry key="location-1" value="BUNDLED:(bundled):Google Checks" />
-        <entry key="location-2" value="LOCAL_FILE:$PROJECT_DIR$/checkstyle.xml:Project Checkstyle" />
+        <entry key="location-0" value="BUNDLED:(bundled):Sun Checks;All" />
+        <entry key="location-1" value="BUNDLED:(bundled):Google Checks;All" />
+        <entry key="location-2" value="LOCAL_FILE:$PROJECT_DIR$/checkstyle.xml:Project Checkstyle;All" />
         <entry key="property-2.config_loc" value="" />
         <entry key="scan-before-checkin" value="false" />
         <entry key="scanscope" value="JavaOnly" />
