+++ src/main/java/org/sheinbergon/dremio/udf/gis/STDistance.java
@@ -21,6 +21,7 @@ import com.dremio.exec.expr.SimpleFunction;
 import com.dremio.exec.expr.annotations.FunctionTemplate;
 import com.dremio.exec.expr.annotations.Output;
 import com.dremio.exec.expr.annotations.Param;
+import org.sheinbergon.dremio.udf.gis.util.FunctionHelpersXL;
 
 @FunctionTemplate(
     name = "ST_Distance",
