+++ src/main/java/org/sheinbergon/dremio/udf/gis/STGeomFromTextSrid.java
@@ -46,10 +46,10 @@ public class STGeomFromTextSrid implements SimpleFunction {
   }
 
   public void eval() {
-    com.esri.core.geometry.ogc.OGCGeometry geom = FunctionHelpersXL.toGeometry(wktInput);
+    com.esri.core.geometry.ogc.OGCGeometry geom = org.sheinbergon.dremio.udf.gis.util.FunctionHelpersXL.toGeometry(wktInput);
     geom.setSpatialReference(com.esri.core.geometry.SpatialReference.create(sridInput.value));
-    byte[] bytes = FunctionHelpersXL.toBinary(geom);
+    byte[] bytes = org.sheinbergon.dremio.udf.gis.util.FunctionHelpersXL.toBinary(geom);
     buffer = buffer.reallocIfNeeded(bytes.length);
-    FunctionHelpersXL.populate(bytes, buffer, binaryOutput);
+    org.sheinbergon.dremio.udf.gis.util.FunctionHelpersXL.populate(bytes, buffer, binaryOutput);
   }
 }
