+++ src/main/java/org/sheinbergon/dremio/udf/gis/STX.java
@@ -22,6 +22,7 @@ import com.dremio.exec.expr.annotations.FunctionTemplate;
 import com.dremio.exec.expr.annotations.Output;
 import com.dremio.exec.expr.annotations.Param;
 import com.esri.core.geometry.ogc.OGCPoint;
+import org.sheinbergon.dremio.udf.gis.util.FunctionHelpersXL;
 
 import javax.annotation.Nullable;
