+++ src/main/java/org/sheinbergon/dremio/udf/gis/util/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Dremio GIS extension common utilities
+ *
+ * @author sheinbergon
+ * @since 0.1.5
+ */
+package org.sheinbergon.dremio.udf.gis.util;
+
